- Nicer approach to 64 bit dcc's
git-svn-id: svn://red.pomac.com/rage@97 55c13d69-e7e2-0310-b639-a4a306287c42

+++ trunk/src/common/dcc.h
@@ -33,7 +33,7 @@ struct DCC
 	int resume_errno;
 
 	GTimeVal lastcpstv, firstcpstv;
-	unsigned int lastcpspos;
+	off_t lastcpspos;
 	int maxcps;
 
 	off_t size;
